uploader/requirements.txt
fastapi
uvicorn
kafka-python
python-multipart  # For handling file uploads
